remove platform dependent tests
/ PlatformAttribute is not present in NUnits netstandard build, but since the tests marked as platform-dependent cannot happen in a core runtime (they apply to old .NET 2 versions), they can be safely ignored
