+++ src/NMoneys.Tests/CurrencyTester.Creation.cs
@@ -47,31 +47,11 @@ namespace NMoneys.Tests
 			Assert.That(Currency.Get(spanish), Is.SameAs(Currency.Eur));
 		}
 
-		[Test, Platform(Include = "Net-2.0")]
-		public void Get_OutdateFrameworkCurrencySymbol_Exception()
-		{
-			CultureInfo bulgarian = CultureInfo.GetCultureInfo("bg-BG");
-			var bulgaria = new RegionInfo(bulgarian.LCID);
-
-			Assert.That(() => Currency.Get(bulgarian), Throws.InstanceOf<InvalidEnumArgumentException>().With.Message.Contains("BGL"), "Framework returns wrong ISOCurrencySymbol (BGL instead of BGN)");
-			Currency lev = Currency.Get(CurrencyIsoCode.BGN);
-
-			Assert.That(lev.IsoSymbol, Is.Not.EqualTo(bulgaria.ISOCurrencySymbol));
-			Assert.That(lev.NativeName, Is.Not.EqualTo(bulgaria.CurrencyNativeName));
-		}
-
 		[Test]
 		public void Get_RegionWithUpdatedInformation_Currency()
 		{
 			var denmark = new RegionInfo("DK");
 			Assert.That(Currency.Get(denmark), Is.SameAs(Currency.Dkk));
-		}
-
-		[Test, Platform(Include = "Net-2.0")]
-		public void Get_RegionWithOutdatedInformation_Exception()
-		{
-			var SerbiaAndMontenegro = new RegionInfo("CS");
-			Assert.That(() => Currency.Get(SerbiaAndMontenegro), Throws.InstanceOf<InvalidEnumArgumentException>().With.Message.Contains("CSD"));
 		}
 
 		[Test]
@@ -97,20 +77,6 @@ namespace NMoneys.Tests
 		public void Get_ObsoleteCurrencyCode_EventRaised(string threeLetterIsoCode)
 		{
 			Action getObsolete = () => Currency.Get(threeLetterIsoCode);
-			Assert.That(getObsolete, Must.Raise.ObsoleteEvent());
-		}
-
-		[Test, Platform(Include = "Net-2.0", Reason = "updated culture uses non-deprecated currency")]
-		public void Get_ObsoleteCulture_EventRaised()
-		{
-			Action getObsolete = () => Currency.Get(CultureInfo.GetCultureInfo("et-EE"));
-			Assert.That(getObsolete, Must.Raise.ObsoleteEvent());
-		}
-
-		[Test, Platform(Include = "Net-2.0", Reason = "updated culture uses non-deprecated currency")]
-		public void Get_ObsoleteRegion_EventRaised()
-		{
-			Action getObsolete = () => Currency.Get(new RegionInfo("EE"));
 			Assert.That(getObsolete, Must.Raise.ObsoleteEvent());
 		}
 
@@ -162,22 +128,6 @@ namespace NMoneys.Tests
 			Assert.That(tried, Is.Null);
 		}
 
-		[Test, Platform(Include = "Net-2.0")]
-		public void TryGet_OutdateFrameworkCurrencySymbol_False()
-		{
-			CultureInfo bulgarian = CultureInfo.GetCultureInfo("bg-BG");
-			var bulgaria = new RegionInfo(bulgarian.LCID);
-
-			Currency lev;
-			Assert.That(Currency.TryGet(bulgarian, out lev), Is.False, "Framework returns wrong ISOCurrencySymbol (BGL instead of BGN)");
-			Assert.That(lev, Is.Null);
-
-			Currency.TryGet(CurrencyIsoCode.BGN, out lev);
-
-			Assert.That(lev.IsoSymbol, Is.Not.EqualTo(bulgaria.ISOCurrencySymbol));
-			Assert.That(lev.NativeName, Is.Not.EqualTo(bulgaria.CurrencyNativeName));
-		}
-
 		[Test]
 		public void TryGet_RegionWithUpdatedInformation_True()
 		{
@@ -186,15 +136,6 @@ namespace NMoneys.Tests
 
 			Assert.That(Currency.TryGet(denmark, out dkk), Is.True);
 			Assert.That(dkk, Is.SameAs(Currency.Dkk));
-		}
-
-		[Test, Platform(Include = "Net-2.0")]
-		public void TryGet_RegionWithOutdatedInformation_False()
-		{
-			var SerbiaAndMontenegro = new RegionInfo("CS");
-			Currency rsd;
-			Assert.That(Currency.TryGet(SerbiaAndMontenegro, out rsd), Is.False);
-			Assert.That(rsd, Is.Null);
 		}
 
 		[Test, TestCaseSource(typeof(Obsolete), nameof(Obsolete.ThreeLetterIsoCodes))]
@@ -210,22 +151,6 @@ namespace NMoneys.Tests
 		{
 			Currency c;
 			Action tryGetObsolete = () => Currency.TryGet(threeLetterIsoCode, out c);
-			Assert.That(tryGetObsolete, Must.Raise.ObsoleteEvent());
-		}
-
-		[Test, Platform(Include = "Net-2.0", Reason = "updated culture uses non-deprecated currency")]
-		public void TryGet_ObsoleteCulture_EventRaised()
-		{
-			Currency c;
-			Action tryGetObsolete = () => Currency.TryGet(CultureInfo.GetCultureInfo("et-EE"), out c);
-			Assert.That(tryGetObsolete, Must.Raise.ObsoleteEvent());
-		}
-
-		[Test, Platform(Include = "Net-2.0", Reason = "updated culture uses non-deprecated currency")]
-		public void TryGet_ObsoleteRegion_EventRaised()
-		{
-			Currency c;
-			Action tryGetObsolete = () => Currency.TryGet(new RegionInfo("EE"), out c);
 			Assert.That(tryGetObsolete, Must.Raise.ObsoleteEvent());
 		}
