+++ src/NMoneys.Tests/NMoneys.Tests.csproj
@@ -53,7 +53,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CurrencyTester.Code.net.cs" />
+    <Compile Include="CurrencyTester.Creation.net.cs" />
     <Compile Include="CustomConstraints\Must.Extensions.net.cs" />
+    <Compile Include="MoneyTester.Creation.net.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\NMoneys\Properties\NMoneys_AssemblyInfo.cs">
       <Link>Properties\NMoneys_AssemblyInfo.cs</Link>
